Changement des types int en size_t dans la structure client de client.h

+++ lib/client.h
@@ -1,8 +1,8 @@
 typedef struct client              //Structure pour la table client
 {
-    int id;
+    size_t id;
     char nom[TAILLE_CHAMP_NOM];
-    int code_postal;
+    size_t code_postal;
     char telephone[10];
     float solde;
     char mot_de_passe[TAILLE_CHAMP_MDP];
